# packages/contracts/openapi/pcbknowledge.openapi.json
{
  "components": {
    "schemas": {
...
      "PrincipalKind": {
        "description": "Authentication actor class stored in the trusted subject mapping.",
        "enum": [
          "HUMAN",
          "SERVICE_ACCOUNT"
        ],
        "title": "PrincipalKind",
        "type": "string"
      },
...
      "Role": {
        "description": "The complete M1 role vocabulary; token-provided roles are ignored.",
        "enum": [
          "DATA_CURATOR",
          "DOMAIN_REVIEWER",
          "KNOWLEDGE_ADMIN",
          "AUDITOR",
          "AGENT_SERVICE"
        ],
        "title": "Role",
        "type": "string"
      },
      "SessionProject": {
        "description": "One explicit project grant from the trusted external-subject mapping.",
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "roles": {
            "items": {
              "$ref": "#/components/schemas/Role"
            },
            "title": "Roles",
            "type": "array"
          }
        },
        "required": [
          "id",
          "roles"
        ],
        "title": "SessionProject",
        "type": "object"
      },
      "SessionResponse": {
        "description": "Trusted current-session projection; it never echoes bearer claims or tokens.",
        "properties": {
          "authenticated_at": {
            "format": "date-time",
            "title": "Authenticated At",
            "type": "string"
          },
          "organization_id": {
            "format": "uuid",
            "title": "Organization Id",
            "type": "string"
          },
          "organization_roles": {
            "items": {
              "$ref": "#/components/schemas/Role"
            },
            "title": "Organization Roles",
            "type": "array"
          },
          "projects": {
            "items": {
              "$ref": "#/components/schemas/SessionProject"
            },
            "title": "Projects",
            "type": "array"
          },
          "subject_id": {
            "format": "uuid",
            "title": "Subject Id",
            "type": "string"
          },
          "subject_kind": {
            "$ref": "#/components/schemas/PrincipalKind"
          }
        },
        "required": [
          "subject_id",
          "subject_kind",
          "organization_id",
          "organization_roles",
          "projects",
          "authenticated_at"
        ],
        "title": "SessionResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "summary": "Evidence-first PCB engineering knowledge service",
    "title": "PcbKnowledge API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/session": {
      "get": {
        "operationId": "get_current_session",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "additionalProperties": true,
                  "description": "Stable base shape for API failures; extension members are permitted.",
                  "properties": {
                    "detail": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ],
                      "title": "Detail"
                    },
                    "instance": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ],
                      "title": "Instance"
                    },
                    "status": {
                      "maximum": 599.0,
                      "minimum": 400.0,
                      "title": "Status",
                      "type": "integer"
                    },
                    "title": {
                      "title": "Title",
                      "type": "string"
                    },
                    "type": {
                      "default": "about:blank",
                      "title": "Type",
                      "type": "string"
                    }
                  },
                  "required": [
                    "title",
                    "status"
                  ],
                  "title": "ProblemDetail",
                  "type": "object"
                }
              }
            },
            "description": "A valid bearer token is required."
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "additionalProperties": true,
                  "description": "Stable base shape for API failures; extension members are permitted.",
                  "properties": {
                    "detail": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ],
                      "title": "Detail"
                    },
                    "instance": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ],
                      "title": "Instance"
                    },
                    "status": {
                      "maximum": 599.0,
                      "minimum": 400.0,
                      "title": "Status",
                      "type": "integer"
                    },
                    "title": {
                      "title": "Title",
                      "type": "string"
                    },
                    "type": {
                      "default": "about:blank",
                      "title": "Type",
                      "type": "string"
                    }
                  },
                  "required": [
                    "title",
                    "status"
                  ],
                  "title": "ProblemDetail",
                  "type": "object"
                }
              }
            },
            "description": "The configured identity provider is unavailable."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Current Session",
        "tags": [
          "identity"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Versioned API gateway",
      "url": "/api/v1"
    }
  ]
}